+++ src/handlers/loadCommands.js
@@ -8,6 +8,11 @@ module.exports = ({
   filesToExclude = [""],
   mainDirectory,
   logger,
+  slashCommandIdsToDelete = [],
+  onSlashCommandsLoading = {
+    delete_global_slash_commands: false,
+    delete_guild_slash_commands: false,
+  },
 }) => {
   const loadCommands = (dir) => {
     const builtInDirectory = dir.includes("../commands/");
@@ -46,6 +51,38 @@ module.exports = ({
       });
     }
   };
+
+  const deleteSlashCommands = async () => {
+    const rest = new REST().setToken(client.config.token);
+    const { clientId, guildId } = client.config;
+    try {
+      for (const id of slashCommandIdsToDelete) {
+        await rest
+          .delete(Routes.applicationCommand(clientId, id))
+          .then(() =>
+            console.log(
+              "[log]",
+              "[Slash CMDs]",
+              `Successfully deleted slash command with ID: ${id}`
+            )
+          )
+          .catch(console.error);
+      }
+    } catch (err) {
+      console.error(
+        "[log]",
+        "[Slash CMDs]",
+        "Error deleting slash commands before loading:",
+        err
+      );
+    }
+  };
+  (async () => {
+    if (slashCommandIdsToDelete.length > 0) {
+      await deleteSlashCommands();
+    }
+  })();
+
   const loadingCommands = directory.map((dirs) => {
     loadCommands(dirs);
   });
@@ -54,80 +91,104 @@ module.exports = ({
     const rest = new REST().setToken(client.config.token);
     const { clientId, guildId, toDeleteSlashCommand } = client.config;
 
-    const slashCommands = await client.commands
-      .map((cmd) => {
-        if (cmd.data) {
-          return cmd.data.toJSON();
-        }
-      })
-      .filter((data) => data !== undefined);
+    const slashCommands = [];
+    client.commands.forEach((cmd) => {
+      if ("data" in cmd && "interactionReply" in cmd) {
+        slashCommands.push(cmd.data.toJSON());
+      }
+    });
 
-    if (slashCommands.length === 0) return;
+    if (slashCommands.length === 0) {
+      console.log("[log]", "[Slash CMDs]", "No slash commands to load.");
+      return;
+    }
 
     try {
-      console.log(
-        `Started refreshing ${slashCommands.length} application (/) commands.`
-      );
       if (guildId) {
+        if (onSlashCommandsLoading.delete_guild_slash_commands === true) {
+          console.log(
+            "[log]",
+            "[Slash CMDs]",
+            "Started deleting guild slash commands..."
+          );
+
+          await rest
+            .put(Routes.applicationGuildCommands(clientId, guildId), {
+              body: [],
+            })
+            .then(() =>
+              console.log(
+                "[log]",
+                "[Slash CMDs]",
+                `Successfully deleted all guild commands before loading new ones.`
+              )
+            )
+            .catch(console.error);
+        }
+
+        console.log(
+          "[log]",
+          "[Slash CMDs]",
+          `Loading ${slashCommands.length} guild slash commands...`
+        );
         const data = await rest.put(
           Routes.applicationGuildCommands(clientId, guildId),
           { body: slashCommands }
         );
 
-        if (typeof toDeleteSlashCommand === "string") {
-          rest
-            .delete(
-              Routes.applicationGuildCommand(
-                clientId,
-                guildId,
-                toDeleteSlashCommand
-              )
-            )
-            .then(() => console.log("Successfully deleted guild command"))
-            .catch(console.error);
-        } else if (toDeleteSlashCommand === true) {
-          rest
-            .put(Routes.applicationGuildCommands(clientId, guildId), {
-              body: [],
-            })
-            .then(() => console.log("Successfully deleted all guild commands."))
-            .catch(console.error);
-        }
-
         console.log(
           "[log]",
-          "[Slash Commands]",
+          "[Slash CMDs]",
           `Successfully Loaded a total of ${data.length} slash commands.`
         );
       } else {
-        const data = await rest.put(Routes.applicationGuildCommands(clientId), {
-          body: slashCommands,
-        });
-        if (typeof toDeleteSlashCommand === "string") {
-          rest
-            .delete(
-              Routes.applicationGuildCommand(clientId, toDeleteSlashCommand)
-            )
-            .then(() => console.log("Successfully deleted guild command"))
-            .catch(console.error);
-        } else if (toDeleteSlashCommand === true) {
-          rest
-            .put(Routes.applicationCommands(clientId), { body: [] })
+        if (
+          onSlashCommandsLoading.delete_global_slash_commands === true ||
+          toDeleteSlashCommand
+        ) {
+          console.log(
+            "[log]",
+            "[Slash CMDs]",
+            "Started deleting global slash commands..."
+          );
+          const deleteAction =
+            toDeleteSlashCommand && typeof toDeleteSlashCommand === "string"
+              ? rest.delete(
+                  Routes.applicationGuildCommand(clientId, toDeleteSlashCommand)
+                )
+              : rest.put(Routes.applicationCommands(clientId), { body: [] });
+
+          deleteAction
             .then(() =>
-              console.log("Successfully deleted all application commands.")
+              console.log(
+                "[log]",
+                "[Slash CMDs]",
+                `Successfully deleted ${
+                  typeof toDeleteSlashCommand === "string"
+                    ? "specific guild command"
+                    : "all application commands"
+                }.`
+              )
             )
             .catch(console.error);
         }
+        console.log(
+          "[log]",
+          "[Slash CMDs]",
+          `Loading ${slashCommands.length} global slash commands...`
+        );
+        const data = await rest.put(Routes.applicationGuildCommands(clientId), {
+          body: slashCommands,
+        });
 
         console.log(
           "[log]",
-          "[Slash Commands]",
+          "[Slash CMDs]",
           `Successfully Loaded a total of ${data.length} Global slash commands.`
         );
       }
     } catch (error) {
-      // And of course, make sure you catch and log any errors!
-      console.error(error);
+      console.error("[Error log]", "[Slash CMDs]", `error`);
     }
   });
 };