+++ src/index.js
@@ -59,6 +59,11 @@ const DiscordFeaturesHandler = async (
       events: false,
       modules: false,
     },
+    slashCommandIdsToDelete = [],
+    onSlashCommandsLoading = {
+      delete_global_slash_commands: false,
+      delete_guild_slash_commands: false,
+    },
     disableUnhandledRejectionHandler = false,
     modulesPreloadTime = 500,
     filesToExcludeInHandlers = {
@@ -134,6 +139,16 @@ const DiscordFeaturesHandler = async (
 
   if (typeof onLoad_list_files.modules !== "boolean") {
     throw new TypeError(`onLoad_list_files.modules must be a boolean value`);
+  }
+
+  if (
+    !Array.isArray(slashCommandIdsToDelete) ||
+    !slashCommandIdsToDelete.every((id) => typeof id === "string")
+  ) {
+    console.warn(
+      "slashCommandIdsToDelete should be an array of strings representing slash command IDs"
+    );
+    slashCommandIdsToDelete = [];
   }
 
   if (typeof process.env.DISCORD_TOKEN === "undefined") {
@@ -217,6 +232,8 @@ const DiscordFeaturesHandler = async (
     filesToExclude: commandsExcluded,
     mainDirectory: directories.main,
     logger: onLoad_list_files.commands,
+    slashCommandIdsToDelete,
+    onSlashCommandsLoading,
   });
   loadEvents({
     client,
@@ -233,7 +250,9 @@ const DiscordFeaturesHandler = async (
 
   client.login(client.config.token).catch((e) => {
     console.warn(e);
-    throw new Error(`Please check if your discord bot token is valid!`);
+    throw new Error(
+      `Please check if your discord bot token is valid! The token should be defined in config.token`
+    );
   });
   await client.wait(modulesPreloadTime);